Stricter warnings for unordered and hash.
There are still warnings in hash_complex_test.

git-svn-id: http://svn.boost.org/svn/boost/trunk@57537 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/unordered/test/unordered/erase_tests.cpp
@@ -65,7 +65,7 @@ void erase_tests1(Container*, test::random_generator generator = test::default_g
         while(size > 0 && !x.empty())
         {
             using namespace std;
-            int index = rand() % x.size();
+            int index = rand() % (int) x.size();
             BOOST_DEDUCED_TYPENAME Container::const_iterator prev, pos, next;
             if(index == 0) {
                 prev = pos = x.begin();
